// packages/sss-token/src/generated/types/roleType.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "gill";

/** Role types supported by the program */
export enum RoleType {
  Minter,
  Burner,
  Pauser,
  Blacklister,
  Seizer,
}

export type RoleTypeArgs = RoleType;

export function getRoleTypeEncoder(): FixedSizeEncoder<RoleTypeArgs> {
  return getEnumEncoder(RoleType);
}

export function getRoleTypeDecoder(): FixedSizeDecoder<RoleType> {
  return getEnumDecoder(RoleType);
}

export function getRoleTypeCodec(): FixedSizeCodec<RoleTypeArgs, RoleType> {
  return combineCodec(getRoleTypeEncoder(), getRoleTypeDecoder());
}